reverted welcome page to just MongoDB, PostgreSQL, and MySQL
