--- client/components/welcome/welcome-modals/welcome-intro.js
@@ -1,33 +1,0 @@
-import React from 'react';
-import RaisedButton from 'material-ui/RaisedButton';
-
-const WelcomeIntro = props => {
-  const styles = {
-    border: '1px solid white',
-    width: '125px',
-    fontSize: '1.2em',
-    color: 'white',
-  };
-  return (
-    <div>
-      <div id='subheading'>Rapidly prototype a full stack React GraphQL Application.</div>
-      <div className='iconContainer'>
-        <img id='icon_graphql' src='./images/graphql.png' />
-        <img id='icon_express' src='./images/express.png' />
-        <img id='icon_react' src='./images/react.png' />
-      </div>
-      <hr className='welcome-hr' />
-      <h4>Select Your Database Type</h4>
-      <div id='buttonsContainer'>
-        <RaisedButton onClick={props.handleWelcomeVersion.bind(null, 'NoSQL', false)} buttonStyle={styles}>
-          NoSQL
-        </RaisedButton>
-        <RaisedButton onClick={props.handleWelcomeVersion.bind(null, 'SQL', false)} buttonStyle={styles}>
-          SQL
-        </RaisedButton>
-      </div>
-    </div>
-  )
-}
-
-export default WelcomeIntro;
