+++ client/components/welcome/welcome.js
@@ -3,35 +3,27 @@ import { connect } from 'react-redux';
 import * as actions from '../../actions/actions.js';
 
 // styling
+import RaisedButton from 'material-ui/RaisedButton';
 import Dialog from 'material-ui/Dialog';
 import './welcome.css';
-
-//components
-import WelcomeIntro from './welcome-modals/welcome-intro.js';
-import WelcomeNoSQL from './welcome-modals/welcome-nosql.js';
-import WelcomeSQL from './welcome-modals/welcome-sql.js';
-import WelcomeJoinMonster from './welcome-modals/welcome-joinmonster.js';
-
 
 const mapStatetoProps = store => ({
   projectReset: store.schema.projectReset,
 });
 
 const mapDispatchToProps = dispatch => ({
-  chooseDatabase: dbName => dispatch(actions.chooseDatabase(dbName)),
+  chooseDatabase: database => dispatch(actions.chooseDatabase(database)),
   handleNewProject: reset => dispatch(actions.handleNewProject(reset)),
 });
 
 class Welcome extends React.Component {
   constructor(props) {
     super(props);
-    this.handleClose = this.handleClose.bind(this);
-    this.handleWelcomeVersion = this.handleWelcomeVersion.bind(this); 
     this.state = {
       open: false,
-      welcomeVersion: 'Intro',
-      welcomeModal: <WelcomeIntro handleWelcomeVersion={this.handleWelcomeVersion} />
     };
+    this.handleClose = this.handleClose.bind(this);
+    this.handleDatabaseClick = this.handleDatabaseClick.bind(this);
   }
 
   componentDidMount() {
@@ -44,30 +36,19 @@ class Welcome extends React.Component {
     this.setState({ open: false });
   }
 
-  handleWelcomeVersion(version, close, SQLType, event){ 
-    if (close){
-      this.props.handleNewProject(false);
-      this.props.chooseDatabase(version)
-      this.setState({
-        open: false,
-        welcomeModal: <WelcomeIntro handleWelcomeVersion={this.handleWelcomeVersion}/>
-      })
-    } else {
-      switch(version) {
-        case 'NoSQL':
-          this.setState({welcomeModal: <WelcomeNoSQL handleWelcomeVersion={this.handleWelcomeVersion}/>});
-          break; 
-        case 'SQL':
-          this.setState({welcomeModal: <WelcomeSQL handleWelcomeVersion={this.handleWelcomeVersion}/>});
-          break; 
-        case 'JoinMonster':
-          this.setState({welcomeModal: <WelcomeJoinMonster SQLType={SQLType} handleWelcomeVersion={this.handleWelcomeVersion}/>});
-          break; 
-      }
-    }
+  handleDatabaseClick(database) {
+    console.log('database', database)
+    this.props.handleNewProject(false);
+    this.props.chooseDatabase(database);
   }
 
   render() {
+    const styles = {
+      border: '1px solid white',
+      width: '125px',
+      fontSize: '1.2em',
+      color: 'white',
+    };
     return (
       <div>
         <Dialog
@@ -78,7 +59,25 @@ class Welcome extends React.Component {
           className='welcome-container'
           paperClassName='welcome-box'
         >
-          {this.state.welcomeModal}
+          <div id='subheading'>Rapidly prototype a full stack React GraphQL Application.</div>
+          <div className='iconContainer'>
+            <img id='icon_graphql' src='./images/graphql.png' />
+            <img id='icon_express' src='./images/express.png' />
+            <img id='icon_react' src='./images/react.png' />
+          </div>
+          <hr className='welcome-hr' />
+          <h4>Select your database type</h4>
+          <div id='buttonsContainer'>
+            <RaisedButton onClick={this.handleDatabaseClick.bind(null, 'MongoDB')} buttonStyle={styles}>
+              MongoDB
+            </RaisedButton>
+            <RaisedButton onClick={this.handleDatabaseClick.bind(null, 'MySQL')} buttonStyle={styles}>
+              MySQL
+            </RaisedButton>
+            <RaisedButton onClick={this.handleDatabaseClick.bind(null, 'PostgreSQL')} buttonStyle={styles}>
+              PostgreSQL
+            </RaisedButton>
+          </div>
         </Dialog>
       </div>
     );
